src/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
gunicorn==20.0.4
gevent==1.4.0
uvicorn
